make community page vertical scroll and change buttons on header
